Added raster callback functions for PWG/Apple Raster input
For input data in PWG/Apple Raster format streaming needs to be supported, meaning that we do not spool the whole job and start processing only then, but instead, we process the data as it comes in, saving memory on very large jobs or allowing for "infinite" jobs where the complete content is not yet determined when the first page is sent.

PDF and PostScript jobs will get spooled as usual, conversion and printing only done on complete submission of the job.

Also done the needed updates for API changes of PAPPL.