Call papplDeviceFlush(device) at the end of the job
In device output filter function ps_print_filter_function() we call
papplDeviceFlush(device) after having sent off all the job's data.
This assures that all data gets to the actual hardware device and does
not stay in a buffer while the device is timing out not getting the
last "showpage" PostScript command and so not printing the last page.