Make 1-bit dithering draft mode also working for images
For image printing the raster output header already needs to be set to
draft mode in the start-job raster callback as PAPPL determines the
pixel value for white padding before the start-page raster callback is
called, leading to black padding when the start-job raster callback
returns with a grayscale mode header while the page is done with a
1-bit black mode header.

For PWG/Apple-Raster jobs we need to switch to the 1-bit dithering
header in the start-page raster callback as the job-options and header
data structures are completely deleted and re-built before every page.

Therefore we do the switching to 1-bit dithering mode in the new
ps_one_bit_dither_on_draft() function and call it from both the
start-job and start-page raster callbacks. Calling it twice does not
break anything.