Added "auto" driver for automatic model selection
Now it is not needed to pick one's printer from a list with ~4000
entries. The first entry in the list is "Automatic Selection" which
automatically selects the model by make/model matching, via the
ps_autoadd() function which is also callback for using the "autoadd"
subcommand of the Printer Application.

ps_autoadd() tries at first to find an exact match of the make and
model fields of the device ID, if this fails it normalizes the device
ID to the same format as the driver names and matches that. As last
mean any PostScript printer ("CMD:POSTSCRIPT;" in device ID) matches
the generic PPD.