Added a test page and a callback function for printing it
PAPPL also supports to send a test page to the printer. An appropriate
callback function to make use of this functionality and the test page
itself got added now.

The test page itself is taken from CUPS, the good old PostScript test
page which came with CUPS 1.0 already 20 years ago. Its big advantage
is that it adapts itself to the actual page size and scales
appropriately. It also shows information about the page geometry and
the PostScript interpreter.

The test page is in color but also prints on monochrome printers, in
grayscale naturally then. So it tests both color reproduction and
color-to-grayscale conversion by the printer.