Added support for the papplJobIsCanceled() function in filter functions
To support the papplJobIsCanceled() function of PAPPL to indicate that
a job has been canceled and its execution should be therefore stopped,
the filter functions of cups-filters got changed to use a callback
function to get note of that the job has stopped. This callback
function support is now used in the PostScript Printer Application to
make the filter functions call papplJobIsCanceled().