If stse file has printer entry with invalid driver, fall back to auto-selection
Re-introduced the mechanism of the driver callback to fall back to
auto-select the printer model and only finally fail if the
auto-selection fails. This prevents a printer from being deleted when
by an update of the Printer Application the PPD file gets removed or
replaced by a one with a different model name.