ps_autoadd(): Added NULL check on driver list and corrected generic driver check
It should never happen that he driver lsit is empty, as then the
Printer Application would not make sense, but we should not segfault
in such a case anyway.

Also, with the "auto" driver being already integral part of PAPPL the
generic driver takes index 0 and not 1 any more in the driver list.