When executing a job, use correct name for the media source PPD option
The standard name for the media source option in PPD files is
"InputSlot", but there are some PPDs where the name is
different. libppd find these odd-named media source options and so we
get it correctly into the driver data, but on job execution
"InputSlot" was hard-coded as option name, making tray selection being
ignored when the name differs. This is corrected here.