Read the PPD file only once and not on every task
Before, the PPD file got read on the setup of the printer (or on the
start of the Printer Application) in the driver setup callback and
then its data structure got immediately freed. After that the PPD got
read again for each job executed.

Now we keep the PPD's data structure in memory and free it only on
shutdown or printer deletion, re-using it for all the jobs. We will
also add moe functions using the PPD data structure in the future.