Added new web interface page to control PostScript-specific functionality
PostScript has some extra functionality to ease the
setup/configuration of the print queue but needs some extra control
points which are not available in the generic Printer Application
framework PAPPL. Therefore we add an extra web admin interface page to
accommodate these additional control points.

The extra functions planned here are

- Selecting which hardware add-ons (like extra trays, duplex unit,
  finishers, ...) are installed on the printer, so that the Printer
  Application advertises only the properties of the actual hardware
  configuration. This is represented by the "Installable Options"
  option group in the PPD file.
- Polling default settings for option settings from the printer to
  ease the initial setup. We are looking into polling mounted media
  ("media-ready") and installed add-ons ("Installable Options")
  separately. Pollable option settings are have PostScript code for
  the polling operation defined in "*?..." attributes assigned to
  each pollable option.

Currently, there is only a non-functional skeleton for an extra web
admin interface page. Next step will be filling it with functionality
and add some helpful text to it.

Also note that not all printers support hardware add-ons or polling of
option settings, so the control points on this page can vary or the
page even not appear for some printers once it gets functional.