ps_driver_setup(): Take into account "Installable Options" settings
Continued on "Installable Options" support - Work in Progress

Now the driver data setup is taking into account the optional hardware
accessories of the printer, as configured via the options in the
"Installable Options" group of the PPD file.

ps_driver_setup() has two modes, controlled by the variable "update":

- Initial setup, which is selected when the driver extension data
  structure is not yet created (update = 0). This loads the PPD file
  and creates all needed data structures
- Update mode, selected if the driver data is set up (update = 1),
  mainly to get in changes on the driver options when the "Installable
  Options" settings have changes.

Now option choices are removed when they do not make sense with the
current accessory configuration (like 2-sided if no duplexer is
installed). When there are less then two choices left, the option is
removed altogether. This affects the "Printing Defaults" page of the
web interface and responses to get-printer-attributes IPP requests.

This is not yet working, appropriate calls to update the driver data
still need to get added.