Added functionality to poll option settings from the printer
PostScript is a full-fledged programming language and many PostScript
printers allow querying settings of options and the presence of
installable hardware accessories executing appropriate PostScript
code. If a setting can get queried, the manufacturer puts the needed
PostScript code into the PPD file, together with the queriable option.

CUPS has already a special command filter, commandtops, to query
settings from a PostScript printer and update the settings of the CUPS
queue appropriately.

This change overtakes the code of CUPS' commandtops into the
PostScript Printer Application so that the printer can get
queried. This way all the experience of developing the original code
gets transferred, especially robustness against broken PostScript code
in the manufacturer's PPD files (see
https://github.com/apple/cups/issues/4028, thanks, Helge Blischke and
Michael Sweet). As there are not the concepts of a side channel or
reading from the printer with timeout, the delay of the printer's
answers has to be handled by reading repeatedly, until a timeout is
reached. All the rest works as in the original.

The queries are triggered on the "Device Settings" page of each
printer, via appropriate buttons. There are two types of queries, one
for the presence of installable hardware accessories (like extra
ppaper trays, duplexer, ...) and one for default settings of job
options (like paper size, resolution, ...). Unfortunately, there is no
way to query which media are loaded in which tray. The Buttons for
querying only appear if the PPD has query code in the appropriate
category of options.

The support for querying the presence of hardware accessories is fully
implemented. Clicking the button queries, saves the results, and
updates the options in the web form appropriately.

Querying of default job option settings is not yet completed. Clicking
the button only shows the query result on the screen, but does not
apply the changes yet.