Completed support for querying job option settings
Now the button "Get printing defaults from the printer" on the "Device
Settings" page of the web admin interface does actually work. Default
values for the PPD options which get polled from the printer are
actually applied to the driver data structure and to the IPP
attributes representing the vendor-specific options.

Needs more testing.

What is missing is visualization of the poll results on the "Device
Settings" page. Planned is to display polled "Installable Options" in
the input form in bold and to add a list of changed job options right
under the button "Get printing defaults from the printer".

Also more explaining test will get added to the "Device Settings"
page.