Show poll results on "Device Settings" web interface page
Now the user can see what came back from the printer when polling the
installable accessory configuration or the option defaults.

When polling the presence of installable accessories the settings
returned by the printer get marked with an asterisk in the web form
now.

As for the option defaults there is no form on the "Device Settings"
page we show a list of option settings which came out from the poll
and turn the button from "Poll" into "Poll again".

In addition, section titles and some explaining text are added to the
page.