Added custom page size support
If the PPD describes support for a custom page size, makes this custom
page size also available in the Printer Application, using the minimum
and maximum dimensions as described in the PPD file.

The 2 media entries for the minimum and maximum dimensions are added
in the beginning of the media list, so that if the PPD file defines
more standard page sizes as they would fit into the maximum number of
slots in PAPPL's data structure (PAPPL_MAX_MEDIA) the custom size
entry is for sure available and so the user can overcome any missing
standard size using the custom size.