Added support for collated copies on PostScript and PDF jobs
The Printer Application advertises to a get-printer-attributes IPP
request:

multiple-document-handling-supported (1setOf keyword) =
   separate-documents-uncollated-copies,separate-documents-collated-copies

This means that one can get collated or uncollated copies using the
"multiple-document-handling" IPP attribute when sending a job.

The setting of this attribute is now passed to the filters when
executing the job. It gets actually made use of only on PostScript and
PDF jobs though, as for images (always single-page jobs) it does not
change anything and for Raster input it is not used by standard.