Added web interface page for uploading extra PPD files
With this functionality we allow users to add their own PPD files to
the Printer Application, for printers not supported by the included
PPD files, especially PPD files with non-free licenses are not
included and so users need to add these.

Currently only the web interface page is added and allows for
selecting PPD files via the facilities of the browser and uploading
them into a reserved directory (last directory in the list of PPD
directories).

This is work in progress and does not perfectly work yet (Hint:
Restart Printer Application to get the uploaded PPDs into driver
list).

Missing yet:

- Driver list needs to be re-created after the PPD upload so that the
  appropriate printer models appear in the driver list in "Add
  Printer" (or get auto-assigned to the printer)

- Feedback of upload: What got successfully uploaded, which
  errors/warning occured?

- List of currently added PPDs with each having a checkbox for
  deleting, delete button, buttons for checking/clearing all
  checkboxes

- Warning if uploaded PPD contains "cupsFilter(2)" line(s)

- Suppressing PPD options where the choices have no PostScript or JCL
  code