Improved driver auto-selection in ps_autoadd() callback function
We do not stop at the first make/model match any more but always go
through the full list, finding the best of all PPDs which match by
make and model.

Highest priority have user-added PPDs, as if a user adds a PPD file
he wants to use it, so auto-selection should prefer it.

After that we prefer English PPDs against any other language. Later,
when PAPPL supports internationalization, we will prefer PPDs in the
user's/the system's language, and if such a PPD is not available,
English.

For the make/model match we prefer if it matches with a device ID
explicitly mentioned in the PPD, if not available we use make and
model from which the driver name is derived.