On "Add PPD files" web interface page report results of upload
When uploading one or more PPD files report which ones could not get
uploaded and which ones actually got uploaded.

Currently PPD files only get rejected if their upload did not complete
(out of disk space or similar) or if there are errors in the PPD file
format or if they are not even a PPD file.

In case of a session ID failure or the value of the submission button
being incorrect, all uploaded PPDs will get removed again.