"Add PPD files" web interface page: Issue warning on PPDs with "*cupsFilter:"
If a PPD file defines a CUPS filter used with it (via a "*cupsFilter:"
or "*cupsFilter2:" line) then it is usually a PPD of a CUPS driver and
not a PostScript PPD. In this case we still accept the file and issue
a warning in the report of uploaded PPDs.

We do not reject the PPD file as sometimes such PPDs still work due to
the fact that some manufacturers introduce a CUPS filter which does
slight changes on PostScript input and passes the PostScript then on
to the printer.  Most options of such a PPD still work without the
filter. This is the case for example with the PPDs for HP PostScript
printers which come with HPLIP.