"Add PPD files" web interface page: List uploaded PPDs and allow to delete them
Now the "Add PPD files" always shows which PPD files are currently in
the directory for user-uploaded PPD files.

In addition there is a facility to delete files to undo accidental
uploads or to remove what is not needed any more. Each listed PPD gets
a checkbox to mark whether one wants to delete it. Clicking the
"Delete" button then does away with the selected files.