Check PPD options whether they have PostScript/PJL code, skip them if not
In the PostScript Printer Application we use PPD files as they were
originally designed: To each option setting a piece of PostScript or
PJL code is assigned and to apply the option setting to a job, to
instruct the printer to apply this setting this code is inserted into
the job's PostScript data stream.

If PPD files have option choices without code, especially more than
one of them in one option, the choices do not make sense, as they do
not do anything different to the job. Only a single choice can be
codeless (like "Auto" for "InputSlot", sending no code for the tray
selection makes the printer choosing by itself) for the option still
to make sense.

Options without code usually only appear in PPD files of CUPS drivers,
where a "*cupsFilter: ..." line defines a filter to pass the job
through. Here the filter applies the codeless options somehow. Most of
these PPDs are for non-PostScript printing though and do not make
sense to be used with this Printer Application, but there are some
which are actually for PostScript printing, but some of the options
are implemented in a CUPS filter (like HP' PostScript PPDs which come
with HPLIP). These PPDs are still usefull, especially if only little
functionality is missing.

Therefore we do not reject PPDs with codeless options but simply skip
these when building the driver data, so that these non-working options
do not appear on the "Printing Defaults" page of the web interface or
in the response to a client's get-printer-attributes request.

Only if the "PageSize" option in the PPD (the only required option in
PPD files) is codeless, we reject the PPD file altogether.