"Add PPD files" web interface page: Check options of uploaded PPDs for code
When uploading a PPD file into the Printer Application, we do not only
check whether it uses a CUPS filter but also whether its option
choices have PostScript or PJL code to be inserted into the job data
assigned.

If there are options without code for two or more choices, the options
are shown as non-working in a warning message (they will also not
appear in the web interface when one actually uses this PPD file). We
do not reject such PPD files altogether as sometimes a PostScript PPD
file uses a filter to implement some, usually non-essential, optiond,
for example passwords for secured printing.

Only if the "PageSize" option does not have code, we reject the PPD
file as this shows for sure that it is not a working PostScript PPD
file.