Do not save the state file while it is still loaded during startup
The status callback ps_status() called papplSystemSaveState() but it
was already called during papplSystemLoadState() during system
startup, making the premature state file saving destroying the state
file. Now we check whether the system is already running, which means
that the state file is already loaded, before saving the state file in
the status callback.

This caused the weird problem reported to PAPPL as
https://github.com/michaelrsweet/pappl/issues/135.