Make the Snap auto-starting the Printer Application
Addad an "apps" entry to snapcraft.yaml with the "daemon: simple"
attribute starting the Printer Application in "server" mode. This way
the Printer Application gets automatically started after installing or
updating the Snap and on every system boot.

The auto-starting Printer Application is running as root, as it is
standard with daemons in Snaps. As libcups puts TLS certificates into
/etc/cups/ssl/ without possibility to change by build time or run time
configuration, we include our own build of libcups in the Snap with
the CUPS_SERVERROOT patched to be /var/snap/ps-printer-app/common.

Note that this prevents from sharing the certificates between all server
processes which need them on the system, for example for enterprise users
to use a CA-signed corporate certificate. Unfortunately, Linux has no
standard certificate storage for the system, so using /etc/cups/ssl/ in
a classic Debian/RPM-based system is the best compromise but does not
work with sandboxed packaging, like Snap.

See also https://github.com/michaelrsweet/pappl/issues/144