Add NULL check for setting vendor options when executing a job
Buggy clients can supply an option (IPP attribute) without
argument. This causes the string content of the IPP attribute to be
NULL which has caused a crash. We check this now and use the default
value then.

This especially happens if the GTK print dialog grabs an IPP printer
directly instead of a CUPS queue and sends the job directly to that
printer, by-passing CUPS.