snapcraft.yaml: Removed unneeded plugs "network-manager" and "log-observe"
To start the development of the PostScript Printer Application the HP
Printer Application got copied and then edited and extended to turn
into the PostScript Printer Application which we have now. Included
was also the snapcraft.yaml. There the hp-printer-app executable is
plugging several interfaces including "network-manager" and
"log-observe".

These two interfaces seem not to be needed when looking through the
code and "log-observe" got actually removed in hp-printer-app. So we
are removing the two now. This especially eases distribution of the
PostScript Printer Application via the Snap Store, as we do not need
to request permission for auto-connection of these interfaces.

"network-manager" is intended to get notified about a change of the
host name, but there is no mention of Network Manager in PAPPL. If it
is really needed for this purpose, probably "network-manager-observe"
would be good enough.