In the Snap patch PAPPL to allow normal-user clients access the Snap's server
The server run by the Snap runs as root, as a Snap is installed into
the system, not into the user's home directory and so a process
started by the Snap installation must be a system process and that is
why Snaps run daemons as root.

Running the Printer Application as a client (from the command line) it
runs as the calling user and PAPPL does not associate it with the
system-wide server which the Snap has started as root. PAPPL always
creates the socket file with the user ID in the file name, so the call
as client uses its own socket file name, does not find the Snap's
server and fires up its own server.

See https://github.com/michaelrsweet/pappl/issues/148

When building the Snap we are patching PAPPL now in a way that if a
client does not find a server uder the user's socket name, it tries
root's socket name (the system-wide server of the Snap in our case),
and only if this also fails it fires up its own server.