Set a default spool directory and allow changing via environment variable
Without default spool directory, PAPPL's default is used and this is a
new subdirectory in the $TMPDIR, with the PID of the ps-printer-app
server in the command line. So with every restart of the server a new
directory gets created, flooding the $TMPDIR with the time.

Now we set a fixed loacation and allow changing it via environment
variable, what we do in the wrapper script for the Snap.