Update README.md because of Snap Store publication
The PostScript Printer Application is available in the Snap Store now
and so one can use it without building it by oneself.

Updated the instructions telling how to download it from the Snap
Store, also added a link to the Snap Store request for auto-connecting
the avahi-control and raw-usb interfaces.