Updated for change of PAPPL's papplPrinterGetDriverAttributes() function
The function papplPrinterGetDriverAttributes() in PAPPL got changed
from returning a pointer to the original driver IPP attributes data
structure to returning a copy of the data structure, letting programs
using this functions stay building without warnings or errors, but not
actually working any more.

This made the installable accessory settings not being saved in the
state file any more (visible defect) and memory leaks being caused
(invisible defect).

This commit adjust everything to correctly work with the new behavior
of the function.