Updated for changes on media verification in PAPPL
PAPPL now denies any media on non-borderless-capable printers which
have narrower margins then the general margins defined in
driver_data->left_right and driver_data->bottom_top, meaning that
these two values describe the minimum margins of all the PPD's page
sizes which are not fully borderless (all 4 margins zero).

Therefore we check the margins of all page sizes which are not fully
borderless and find the minimum values for each of the four sides,
then we take the minimum of the resulting minima for left and right
for driver_data->left_right and of the resulting minima for top and
bottom for driver_data->bottom_top.