Handle boolean PPD options correctly when printing a job
When under the vendor-specific PPD options is a boolean option
(declared as boolean type in the PPD, like "ManualFeed" for the HP
LaserJet 4000) and it gets registered as boolean IPP attribute (when
human-readable choice names are "True" and "False") the handling of
the options when printing a job caused a crash.

In this case one can set up the printer with the PPD and set all the
defaults for installable accessories, media, and printing defaults but
one cannot actually print. And when running in a Snap the automatic
restart of the Printer Application by systemd even hides the crash.