PAPPL added support for Printer Application as system service
In PAPPL support for the Printer Application as system service is
added now, especially if the Printer Application runs a root and users
run the Printer Application as client (command line interface) the
commands are applied to the Printer Application running as root and no
new personal server process for the user is created.

This functionality was provided by a patch in this Snap before, this
patch is removed now.

See https://github.com/michaelrsweet/pappl/issues/148

Also handling of directories got improved and therefore additional
arguments were added to the ./configure command line of PAPPL.

See https://github.com/michaelrsweet/pappl/issues/128