Fixed mDNS host name (".local" names) look-up from within the Snap
The Core Snaps core18 and core20 on which the Snaps are based do not
support mDNS host name look-up, so if a device in the network assigns
itself an mDNS (ZeroConf) host name (a name with ".local" domain) the
snapped application is not able to look up this device's IP address
and so not able to access the device.

The problem should be fixed in the Core Snaps but for the time being
there is a workaround to be performed in the application Snaps.

One needs:

1. /etc/nsswitch.conf with "hosts: files mdns4_minimal [NOTFOUND=return] dns"

For this we install the libc-bin package in a separate empty part as
stage package as it contains the original nsswitch.conf as used by the
Core Snaps.  We edit this file with Perl to add the mDNS support and
then prime only this file.

With a Snap layout we mirror the file into /etc/nsswitch.conf. This
affects only the Snap itself, so no danger of breaking the host system
and no special permission needed.

2. Install libnss-mdns*.so*

We install the libraries which implement the lookup by installing
libnss-mdns as stage package in a separate empty part.

This is a known problem. See

https://bugs.launchpad.net/bugs/1838038

Launchpad bug report, originally on the Chromium Browser Snap but
extended to the Core Snaps as the bug is actually there.

https://forum.snapcraft.io/t/no-mdns-support-in-snaps-should-core-have-a-modified-nsswitch-conf/

https://forum.snapcraft.io/t/how-can-i-lookup-host-name-from-snap-core-using-avahi-mdns/

Discussion of the problem on the snapcraft.io forum

https://forum.snapcraft.io/t/install-single-file-from-stage-packages/

How to install a single file from a package (in our case for
nsswitch.conf from libc-bin