When setting up a new printer use default paper size of user's location
When the user sets up a new printer and the PPD file for it has Letter
or A4 as default for the PageSize option, THe default page size is
determined by the user's location (Letter for US and Canada, A4 for
the rest of the world) using the environment variables of the user's
locale setting (LC_PAPER, LC_CTYPE, LC_ALL, LANG).

PPD files have Letter as default in most cases but A4 is the standard
page size at most places on the world, so for most people printer
setup gets easier now.