Refresh default media collection when changing ready media
When changing the loaded media on the "Media" tab in the web interface
and clicking "Save Changes", The changes in the tray selected as
default are not transfered to the default media collection, so that if
one prints immediately afterwards, the old settings, taken from the
default media collection, are used for the job. One has to go to the
"Printing Defaults" tab and click "Save Changes" there to refresh the
default media collection with the data from the selected tray.

This commit corrects the issue by also refreshing the default media
collection when clicking "Save Changes" on the "Media" tab.

This is a bug in PAPPL, reported as

https://github.com/michaelrsweet/pappl/issues/164

with a proposed patch.

This commit applies the patch to the PAPPL used by this Snap to fix
the bug for the time being until it get fixed in PAPPL upstream.