Added support for CUPS filters defined in the PPD files
PPD files in Adobe's original design were only for PostScript printers
and did not define filter or driver executables, but only how to
modify the PostScript data stream for the specifics of each PostScript
printer model, especially to convey the option settings which the user
has supplied with a job. This way they were
operating-system-independent.

Later on, printer manufacturers made OS-specific PPD files to allow
for more complex manipulations of the PostScript code and user options
other than enumerated choice and boolean, especially strings for
passwords and fax numbers. Under Linux/Unix with CUPS they opted for
adding filter executables for these operations, defined with
"*cupsFilter:" or "*cupsFilter2:" lines in the PPD, as drivers for
non-PostScript printers, but the filter takes PostScript as input,
modifies it, and passing it on the PostScript.

Previous versions of the PostScript Printer Application do not support
this. They ignored the filter and did not list options which do not
simply insert PostScript (or JCL/PJL) code into the output stream. One
could still print, but some, usually less commonly used options were
not available (secure, PIN-based printing in most cases).

This commit adds support for such filters. If the filter specified in
the PPD file is available (in the directory specified by the
FILTER_DIR environment variable or in /usr/lib/ps-printer-app/filter)
it will be made use of and all options get listed, if not, we get the
same behavior as before. The filter gets called right before sending
the data off to the printer, after option setting code already
inserted into the PostScript data stream, same as if the PPD and the
filter get used by a CUPS queue.

Note that for making fully use of this feature, we still need to add
support for CUPS' PPD extension for custom PPD option settings
(arbitrary strings, numbers, ...) to this Printer Application.

This is also a step further to support general retro-fitting of CUPS
drivers, but note that currently only filters which take PostScript as
input are supported, not those taking Raster or PDF.