Added the "hpps" filter of HPLIP to the Snap
The PPD files for HP PostScript printers included in the Snap of the
PostScript Printer Application come from HP's HPLIP package. These PPD
files use a simple CUPS filter names "hpps" to implement PIN-protected
printing of private jobs (also known as "Locked Print"). Using the
PPDs without the filter (simply ignoring the "*cupsFilter: ..." entry
in the PPD file) makes the feature not working.

As the options for printing private jobs are implemented in the "hpps"
filter and not in the printer itself, they do not supply PostScript
code to be inserted into the PostScript output stream and therefore
were not shown in the web interface of previous versions of the
PostScript printer appication.  They did not make sense without the
"hpps" filter.

Now, as since the previous commit we support CUPS filters in
PostScript PPD files, we are adding the "hpps" filter to the Snap and
so it gets detected and used by the Printer Applications for the HP
PPD files which request it. With this also all options get shown in
the web interface.

Especially the options for printing private jobs appear,
"secure-printing" to turn on and off the functionality and one option
for each digit of the PIN. With the function activated the printer
does not print the job right away but holds it in its internal
storage. To print the job, the user has to select it from the "Private
Jobs" list on the printer's front panel and enter the same PIN as he
has supplied with the job when sending it to the printer.

This way HP's PostScript printers are fully supported by the
PostScript Printer Application Snap now.

Note that we cannot support PIN-based printing (Locked Print/Document
Server/User Code) on printers from Ricoh and OEM now. We could add the
needed "foomatic-rip" filter from cups-filters, but for entering the
PIN there text/string options (a CUPS extension for PPDs) are used
which are not yet supported by PAPPL. See the feature request with
patches to solve this:

https://github.com/michaelrsweet/pappl/issues/142