Added options to the filter call which CUPS also adds to its filter calls
When CUPS calls the filters to execute a job it does not only add the
PPD options but also some extra options which some filters make use
of. This commit makes several of them also be added to the call of
filter functions and PPD-defined CUPS filters when a job is executed
in the Printer Application.

As most filter functions are derived from CUPS filters and some PPD
files request a call of an additional CUPS filter, we add these
options to get a maximum from the functionality of filter functions
and filters.

The options we add are:

- job-uuid
- job-originating-user-name
- job-originating-host-name
- time-at-creation
- time-at-processing

The options get only added if we find values for them.

Especially the "hpps" filter from HPLIP which is used by the
PostScript PPD files for HP printers (and included in the Snap of the
PostScript Printer Application) uses job-uuid for printer-internal job
accounting. So with this we get full functionality for HP's PostScript
PPD files.