Add foomatic-rip for Ricoh printers and support for string options
Ricoh printers use foomatic-rip as CUPS filter to support options for
PIN-based secure printing (Locked Print, Document Server, User
Code). This is for converting string-type option settings into
appropriate PostScript code.

The string-type options for input of the PINs are vendor options in
the Printer Application and in PAPPL this option type clashes with
hidden options for storing individual data in the state file. For this
a bug is reported:

https://github.com/michaelrsweet/pappl/issues/142

A solution is suggested and attached to the bug as a patch but not yet
accepted into PAPPL. In this commit of the PostScript Printer
Application the patch is applied to the included PAPPL to allow full
support for {IN-secured printing on Ricoh (and OEM) printers.

In addition, a patch on the Printer Application itself is applied
(also attached to the PAPPL bug report) to once use the new method of
the patched PAPPL for saving individual data in the state file and
second, support the CUPS PPD extension for custom option values:

https://openprinting.github.io/cups/doc/spec-ppd.html#OPTIONS

With all this, if PPDs allow input of custom values in string or
numeric form to some options using the CUPS extension, appropriate
input fields appear as additional options under the vendor options on
the "Printing Defaults" page of the web interface.

This especially happens for the PIN options of the Ricoh (and OEM)
printers and so the secure printing functionality is available for
users of the PostScript Printer Applications.

This is also in preparation to further Printer Applications
retro-fitting classic CUPS drivers, especially for Foomatic which
contains several options using CUPS' PPD extension for accepting
custom values.

The patches will get removed as soon as the PAPPL bug is fixed.