Support for CUPS' PPD extension to support custom option values
CUPS has an extension for PPD files so that instead of having only
options which allow the user to select from enumerated choices one can
also select a "Custom" choice and then enter string or numeric values
freely. Here is the documentation of this CUPS extension:

https://openprinting.github.io/cups/doc/spec-ppd.html#OPTIONS

As we want to retro-fit classic CUPS drivers and PPD files into
Printer Applications without needing to do major changes on the
original drivers and PPDs (we do not have the printers to test our
retro-fit), we add support for this extension also to our Printer
Application code with this commit.

Due to limitations of PAPPL floating point numbers of all kinds as
custom paramaters are not supported. In such a case the possibility of
entering a custom value for an option will not get provided. But such
cases are rare.

In the PostScript Printer Application it primarily provides a part of
what is needed to support PIN-based secure printing on Ricoh (and OEM)
printers. Here the options to enter the PINs are using the CUPS
extension to allow entering individual, user-selected PINs.

Note that this does not work around PAPPL's problem of not supporting
string/text-style vendor options:

https://github.com/michaelrsweet/pappl/issues/142

Therefore we still need patches for building the Snap so that it
supports PIN-based printing on Ricoh printers.

This commit differs from the patch we applied to the Snap from the
previous commit (953b3fcd) as follows:

- The changes needed on the Printer Application because of the patch
  to fix PAPPL are not contained, so that our Printer Application
  continues to build and work with the current state of the PAPPL
  code.
- The custom option setting support was tested with integer options
  now and fixes applied to make these options work.
- If custom options have parameters of unsupported types (for example
  floating point numbers) the "Custom" setting is suppressed and only
  enumerated choices can get selected.
- Guards agains buggy PPDs are added. As for the vendor options we
  generate the IPP attribute names from the human-readable strings, we
  skip options if they have a human-readable name already used by
  another option and we skip choices if their human-readable strings
  got already used by other choices of this option.
- Fixed support for PPD files without "InputSlot" option. In this case
  we create a single media source named "default" and create a single
  media-ready entry for it. The fix makes sure that it also gets
  created when starting the Printer Application and updating the
  driver data for installable accessories. Before, we had no
  media-ready entry at all in such case which ended up with an
  incomplete option list on the "Printing Defaults" page and later (on
  shutdown for sure) a crash.