Updated patches for building the Snap with string/text-option fix on PAPPL
With the previous commit support for CUPS' PPD extension for custom
option settings got added to the code base of the PostScript Printer
Application.

As this functionality was provided in a patch for the Snap build
before, we have to chenge the patch on the PostScript Printer
Application to only contain the adaptations to work together with
PAPPL patched to get support for string/text-style vendor options
fixed.